pyproject.toml
[tool.coverage.run]
omit = "*/tests/*"
plugins = ["django_coverage_plugin"]

[tool.django-stubs]
django_settings_module = "jotlet.settings"

[tool.djlint]
custom_blocks = "element,slot,setvar"
format_attribute_template_tags = true
format_css = false
format_js = false
ignore = "H016"
max_line_length = 118
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.mypy]
disable_error_code = ["attr-defined"]
exclude = ["migrations"]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "allauth.*",
    "axes.*",
    "cacheops.*",
    "channels.*",
    "crispy_bootstrap5.*",
    "crispy_forms.*",
    "debug_toolbar",
    "django_cleanup.*",
    "django_redis",
    "environ",
    "huey_monitor.*",
    "pytest_lazy_fixtures.*",
    "sorl.*",
    "storages.backends.*",
    "tree_queries.*",
]

[tool.poetry]
name = "jotlet"
version = "1.11.4"
description = ""
authors = ["Oscar Siles Brügge <siles.oscar@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
boto3 = "^1.28.76"
channels = "^4.0.0"
channels-redis = "^4.1.0"
crispy-bootstrap5 = "^2023.10"
Django = { extras = ["argon2"], version = "^4.2.7" }
django-allauth = "0.58.1"
django-anymail = { extras = ["mailgun", "sendgrid"], version = "^10.2" }
django-auto-prefetch = "^1.8.0"
django-axes = { extras = ["ipware"], version = "^6.1.1" }
django-cacheops = "^7.0.2"
django-cleanup = "^8.0.0"
django-cors-headers = "^4.3.0"
django-csp = "^3.7"
django-environ = "0.11.2"
django-extra-views = "^0.14.0"
django-filter = "^23.3"
django-htmx = "^1.17.0"
django-huey-email-backend = "^0.1.1"
django-huey-monitor = "0.7.1"
django-redis = "^5.4.0"
django-storages = "^1.14.2"
django-tree-queries = "^0.15.0"
gunicorn = "^21.2.0"
httpx = "0.25.1"
huey = "^2.5.0"
Pillow = "^10.1.0"
psycopg = { extras = ["binary"], version = "^3.1.12" }
python = ">=3.10,<4.0.0"
redis = { extras = ["hiredis"], version = "^5.0.1" }
sorl-thumbnail = "^12.10.0"
tomli = "^2.0.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
whitenoise = { extras = ["brotli"], version = "^6.6.0" }

[tool.poetry.group.dev.dependencies]
coverage = { extras = ["toml"], version = "^7.3.2" }
django-coverage-plugin = "^3.1.0"
django-debug-toolbar = "^4.2.0"
django-debug-toolbar-template-profiler = "^2.1.0"
django-filter-stubs = "^0.1.3"
django-stubs = { extras = ["compatible-mypy"], version = "^4.2.6" }
djlint = "^1.34.0"
huey-stubs = "^0.0.2"
pytest-cov = "^4.1.0"
ruff = "^0.1.4"
types-factory-boy = "^0.4.1"
types-pillow = "^10.1.0.0"

[tool.poetry.group.test.dependencies]
daphne = "^4.0.0"
factory-boy = "^3.3.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-django = "^4.6.0"
pytest-factoryboy = "^2.6.0"
pytest-lazy-fixtures = "^1.0.1"
pytest-sugar = "^0.9.7"
pytest-xdist = "^3.3.1"

[tool.pyright]
exclude = ["**/migrations/*.py"]
reportGeneralTypeIssues = false
reportMissingTypeStubs = false
reportOptionalMemberAccess = false
reportPrivateUsage = false
reportShadowedImports = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=jotlet --cov-report xml:cov.xml --cov-report html:htmlcov"

[tool.ruff]
exclude = ["manage.py", "**/migrations/*.py"]
line-length = 118
target-version = "py312"

[tool.ruff.format]
indent-style = "space"
quote-style = "double"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN",
    "EXE",
    "T10",
    "TD",
    # Individual Rules
    "ARG002",
    "ARG003",
    "COM812",
    "COM819",
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D203",
    "D205",
    "D212",
    "D400",
    "D401",
    "D415",
    "D300",
    "E111",
    "E114",
    "E117",
    "ISC001",
    "ISC002",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "RUF012",
    "FIX002",
    "TRY003",
    "W191",
]

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.flake8-self]
ignore-names = ["_meta"]

[tool.ruff.per-file-ignores]
"**/management/commands/*.py" = ["A003"]
"**/models.py" = ["A003"]
"**/signals.py" = ["ARG001"]
"**/tests/*.py" = ["PLR0913", "S101", "SLF001"]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
